refactor: 코드 리뷰 피드백 반영 - 성능 최적화 및 보안 강화
주요 변경사항:
1. ComingSoonPage 컴포넌트 분할 (694줄 → 6개 섹션별 컴포넌트)
   - HeroSection, ProblemSection, SolutionSection
   - HowItWorksSection, VisionSection, FinalCTASection, Footer

2. React 성능 최적화 적용
   - 모든 섹션 컴포넌트에 React.memo 적용
   - useCallback을 통한 이벤트 핸들러 메모이제이션
   - displayName 추가로 디버깅 지원

3. Rate Limiting 시스템 개선
   - Redis 기반 rate limiter 유틸리티 생성
   - 메모리 기반에서 Redis 준비된 구조로 전환
   - 향후 Redis 연동 준비

4. 이메일 유효성 검증 강화
   - RFC 5322 준수 정규표현식 적용
   - 이메일 길이 제한 및 형식 검증 강화

5. API 보안 및 CORS 개선
   - CORS 헤더 적절히 설정
   - OPTIONS 요청 preflight 지원
   - Rate limit 헤더 응답에 포함

6. IP 주소 추출 로직 보완
   - 프록시/로드밸런서 환경 고려
   - Cloudflare, Vercel 등 플랫폼별 헤더 지원
   - RFC 7239 Forwarded 헤더 파싱

7. 상세한 서버사이드 로깅 시스템
   - 구조화된 JSON 로그 형식
   - 성능 메트릭 (DB 쿼리 시간, 총 처리 시간)
   - 개인정보 보호를 위한 이메일 마스킹
   - 오류 추적 및 디버깅 정보

기술적 개선:
- TypeScript 타입 안전성 강화
- 에러 핸들링 및 로깅 표준화
- 보안 헤더 및 CORS 정책 적용
- 프로덕션 환경 최적화
